id값 Integer인것들 전부 Long으로 변경

+++ src/main/java/com/example/ormi5projectteam4/controller/rest_controller/AdminController.java
@@ -71,7 +71,7 @@ public class AdminController {
     }
 
     @PutMapping("/post/{id}")
-    public ResponseEntity<Post> changePostApproveStatus(@PathVariable Integer id, @RequestParam ApproveStatus approveStatus){
+    public ResponseEntity<Post> changePostApproveStatus(@PathVariable Long id, @RequestParam ApproveStatus approveStatus){
         Post updatePost = adminService.changePostApproveStatus(id, approveStatus);
         if(updatePost == null){
             return ResponseEntity.notFound().build();
